Added LazyReader module.
This is essentially the same as Reader. The difference is that the
function 'lazy' creates curried functions which are also lazily
evaluated and LazyReader deals with those lazy functions.
Also added 'force' method and function to Container.py
and modified Monad.py to work correctly with lazy of regular curried
functions.